+++ src/main/java/seedu/mealcompanion/command/factory/misc/RecipeNeedCommandFactory.java
@@ -3,7 +3,7 @@ package seedu.mealcompanion.command.factory.misc;
 import seedu.mealcompanion.MealCompanionSession;
 import seedu.mealcompanion.command.ExecutableCommand;
 import seedu.mealcompanion.command.factory.ExecutableCommandFactory;
-import seedu.mealcompanion.command.misc.RecipeNeedCommand;
+import seedu.mealcompanion.command.recipe.RecipeNeedCommand;
 import seedu.mealcompanion.parser.CommandArguments;
