+++ src/main/java/seedu/mealcompanion/command/factory/misc/RecipePossibleCommandFactory.java
@@ -3,7 +3,7 @@ package seedu.mealcompanion.command.factory.misc;
 import seedu.mealcompanion.MealCompanionSession;
 import seedu.mealcompanion.command.ExecutableCommand;
 import seedu.mealcompanion.command.factory.ExecutableCommandFactory;
-import seedu.mealcompanion.command.misc.RecipePossibleCommand;
+import seedu.mealcompanion.command.recipe.RecipePossibleCommand;
 import seedu.mealcompanion.parser.CommandArguments;
 
 //@@author ngyida
