+++ src/main/java/seedu/mealcompanion/command/factory/misc/RecipeRandomCommandFactory.java
@@ -3,7 +3,7 @@ package seedu.mealcompanion.command.factory.misc;
 import seedu.mealcompanion.MealCompanionSession;
 import seedu.mealcompanion.command.ExecutableCommand;
 import seedu.mealcompanion.command.factory.ExecutableCommandFactory;
-import seedu.mealcompanion.command.misc.RecipeRandomCommand;
+import seedu.mealcompanion.command.recipe.RecipeRandomCommand;
 import seedu.mealcompanion.parser.CommandArguments;
 
 //@@author ngyida
