+++ src/main/java/seedu/mealcompanion/command/ingredients/ClearCommand.java
@@ -1,4 +1,4 @@
-package seedu.mealcompanion.command.misc;
+package seedu.mealcompanion.command.ingredients;
 
 import seedu.mealcompanion.MealCompanionSession;
 import seedu.mealcompanion.command.ExecutableCommand;
