+++ src/main/java/seedu/mealcompanion/command/recipe/RecipeAlmostCommand.java
@@ -1,4 +1,4 @@
-package seedu.mealcompanion.command.misc;
+package seedu.mealcompanion.command.recipe;
 
 import seedu.mealcompanion.MealCompanionSession;
 import seedu.mealcompanion.ingredient.Ingredient;
