+++ src/main/java/seedu/mealcompanion/command/recipe/RecipeCommand.java
@@ -1,4 +1,4 @@
-package seedu.mealcompanion.command.misc;
+package seedu.mealcompanion.command.recipe;
 
 import seedu.mealcompanion.command.ExecutableCommand;
 import seedu.mealcompanion.ingredient.Ingredient;
